requirements.txt
requests==2.25.1
folium==0.11.0
pandas==1.1.4
streamlit==0.78.0
cartoframes==1.1.1
geopandas==0.9.0
geopy==2.1.0
haversine==2.3.0
Pillow==8.1.2
pymongo==3.11.3
python-dotenv==0.15.0
Shapely==1.7.1
